MemberProviderで処理するように修正
