+++ src/Eccube/Security/Core/User/MemberProvider.php
@@ -1,0 +1,80 @@
+<?php
+
+namespace Eccube\Security\Core\User;
+
+use Eccube\Entity\Master\Work;
+use Eccube\Entity\Member;
+use Eccube\Repository\MemberRepository;
+use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
+use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
+use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
+
+class MemberProvider implements UserProviderInterface
+{
+    /**
+     * @var MemberRepository
+     */
+    protected $memberRepository;
+
+    public function __construct(MemberRepository $memberRepository)
+    {
+        $this->memberRepository = $memberRepository;
+    }
+
+    /**
+     * Loads the user for the given username.
+     *
+     * This method must throw UsernameNotFoundException if the user is not
+     * found.
+     *
+     * @param string $username The username
+     *
+     * @return UserInterface
+     *
+     * @throws UsernameNotFoundException if the user is not found
+     */
+    public function loadUserByUsername($username)
+    {
+        $Member = $this->memberRepository->findOneBy(['login_id' => $username, 'Work' => Work::WORK_ACTIVE_ID]);
+
+        if (!$Member) {
+            throw new UsernameNotFoundException(sprintf('Username "%s" does not exist.', $username));
+        }
+
+        return $Member;
+    }
+
+    /**
+     * Refreshes the user.
+     *
+     * It is up to the implementation to decide if the user data should be
+     * totally reloaded (e.g. from the database), or if the UserInterface
+     * object can just be merged into some internal array of users / identity
+     * map.
+     *
+     * @return UserInterface
+     *
+     * @throws UnsupportedUserException if the user is not supported
+     */
+    public function refreshUser(UserInterface $user)
+    {
+        if (!$user instanceof Member) {
+            throw new UnsupportedUserException(sprintf('Instances of "%s" are not supported.', get_class($user)));
+        }
+
+        return $this->loadUserByUsername($user->getUsername());
+    }
+
+    /**
+     * Whether this provider supports the given user class.
+     *
+     * @param string $class
+     *
+     * @return bool
+     */
+    public function supportsClass($class)
+    {
+        return Member::class === $class;
+    }
+}
